PR Cleanup, more `MockClient` unit tests (#130)
* Fixed an issue with the `EqRequestMatcher`
* Adds `expectResponse` helper functions
* Adds unit tests around multiple requests being registered with `MockClient`
