--- Source/Testing/Matchers/RequestMatcher.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-internal protocol RequestMatcher {
-  func match<T: Request>(_ request: T) -> Result<T.ResponseObject, SwishError>?
-}
