+++ Swish.xcodeproj/Swoosh_Info.plist
